% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAdjacencyMatrix.R
\name{calculateAdjacencyMatrix}
\alias{calculateAdjacencyMatrix}
\title{\code{calculateAdjacencyMatrix}}
\usage{
calculateAdjacencyMatrix(candidate_network, pf_data)
}
\arguments{
\item{candidate_network}{A \code{PrixFixeNetwork} generated from \code{constructCandidateNetwork}}

\item{pf_data}{A \code{PFData} object generated from \code{PFDataLoader}}
}
\value{
\code{candidate_network} with updated \code{adjancency_matrix} slot
}
\description{
Parallel calculate the adjacency matrix of a candidate network given the
co-function network. Called within \code{constructCandidateNetwork} and not
intended to be called by user directly.
}
