% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGeneScore.R
\name{calculateGeneScores}
\alias{calculateGeneScores}
\title{Calculate the Prix Fixe gene score}
\usage{
calculateGeneScores(network, pf_data)
}
\arguments{
\item{network}{A \code{PrixFixeNetwork}. Probably selected using \code{getTopNetwork}.}

\item{pf_data}{A \code{PFData} object.}
}
\description{
Evaluate the gene contributions to a given subnetwork.
}
\section{Scoring Methodology}{

To score a gene contribution to a subnetwork we first calculate the null locus
scenario in which we remove, or zero out, a given locus and calculate the
density of the resulting network. Next, we iteratively consider all genes
associated with that locus by substituting and caculating the density. The
difference between each gene density and the empty locus density is the score
for that gene. If the gene has no connectivity to the subnetwork, then we give
it an \code{NA} value.

To implement this method we simplify the problem. First, we are only considering
one locus at a time and therefor all other loci remain constant and can be
ignored here. Due to this, the null locus density must be zero because there
is no connectivity by definition. Therefore, the only value we need to derive
is the sum of the connections between the permuted gene and the other fixed
loci.

Moreover, to quickly calculate the contributions we create a copy of a network
and modify the genes to include the network genes and all genes associated
with a given locus. The adjacency matrix is then recalculated, and filtered
such that the rows only contain the locus genes and the columns are the network
genes. The contribution for each one of the locus genes is then calculated
as the corresponding row sum.
}

