% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileLoci.R
\name{compileLoci}
\alias{compileLoci}
\title{\code{compileLoci}}
\usage{
compileLoci(loci_data)
}
\arguments{
\item{loci_data}{A tibble containing the raw loci information provided by \code{loadLociData}.}
}
\value{
A list of \code{Locus} objects
}
\description{
Reformat the loci data from tibble into a list of \code{Locus} objects. Called
within \code{PFDataLoader} and not intended to be called by user.
}
