% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructCandidateNetwork.R
\name{constructCandidateNetwork}
\alias{constructCandidateNetwork}
\title{\code{constructCandidateNetwork}}
\usage{
constructCandidateNetwork(pf_data, member = c("null_members", "true_members"))
}
\arguments{
\item{pf_data}{a \code{PFData} object generated by \code{PFDataLoader}}

\item{member}{a character string of either \code{true_members} or \code{null_members}.}
}
\value{
A \code{PrixFixeNetwork}
}
\description{
Create a \code{PrixFixeNetwork} by randomly sampling genes from loci. Called within
\code{initializePopulation} and not intended to be called by user.
}
