% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluatePopulationSignificance.R
\name{evaluatePopulationSignficance}
\alias{evaluatePopulationSignficance}
\title{Determine the significance of a population of PrixFixeNetworks}
\usage{
evaluatePopulationSignficance(pf_data, population, num_trials = 10)
}
\arguments{
\item{pf_data}{a \code{PFData} object generated by \code{PFDataLoader}}

\item{population}{a list of \code{PrixFixeNetwork} objects}

\item{num_trials}{The number of random trials to generate the null distribution}
}
\value{
a p-value
}
\description{
Given a population (i.e., list) of PrixFixeNetworks calculate the mean
network density across the population. Compare the mean density to a distribution
of random network densities.
}
\section{Calculating the null model distribtion}{

In order to calculate the null model we want to sample genes from the total
co-function network to generate random loci. In order to mitigate problems
with node-degree effects, we first categorically bin the cofunction network
and then replace original loci genes with a random choice from the same bin.
}

\section{Calculating the p-value}{

The p-value here is defined as the fraction of null model densities that are
greater than the average true network density.
}

\examples{
\dontrun{
# load example PFData (FA genes)
data(PF_FanconiAnemia)
# generate population of subnetworks
population <- initializePopulation(PF_FanconiAnemia, population_size=100, "true_members")
# evaluate population significance
population_significance <- evaluatePopulationSignificance(pf_data, population, 100)
}

}
