% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDegreeInfo.R
\name{getDegreeInfo}
\alias{getDegreeInfo}
\title{\code{getDegreeInfo}}
\usage{
getDegreeInfo(cfn_data, bin = TRUE, n_bins = 128)
}
\arguments{
\item{cfn_data}{A tibble containing co-function network data generated by \code{loadCFNData}.}

\item{bin}{Boolean on whether to bin the degrees for each gene. Default is \code{TRUE}.}

\item{n_bins}{Number of bins to separate data into if \code{bin=TRUE}.}
}
\value{
A formatted tibble with degree information
}
\description{
Calculate the connectivity degrees for the co-function network. Called by
\code{PFDataLoader}, not intended to be directly invoked by user.
}
