% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkDensity.R
\name{getNetworkDensity}
\alias{getNetworkDensity}
\title{\code{getNetworkDensity}}
\usage{
getNetworkDensity(population, return_mean = TRUE)
}
\arguments{
\item{population}{a list of \code{PrixFixeNetwork} objects}

\item{return_mean}{Boolean of whether to return the mean or the vector of network densities. Default is TRUE.}
}
\value{
Mean density value if \code{return_mean=TRUE}, else a vector of density values
}
\description{
Worker function to extract density information for a population of PrixFixeNetworks.
Not intended to be called by user. Called by \code{evaluePopulationSignificance}.
}
