% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopNetwork.R
\name{getTopNetwork}
\alias{getTopNetwork}
\title{Extract the top network from a population}
\usage{
getTopNetwork(population, multiple = FALSE)
}
\arguments{
\item{population}{a list of \code{PrixFixeNetwork} objects}

\item{multiple}{a Boolean on whether to return multiple top networks. Currently not functional.}
}
\value{
a \code{PrixFixeNetwork} object
}
\description{
Identify and return the top network (as determined by the density) from a list
of \code{PrixFixeNetwork} objects.
}
\examples{
\dontrun{
# load example PFData (FA genes)
data(PF_FanconiAnemia)
# generate population of subnetworks
population <- initializePopulation(PF_FanconiAnemia, population_size=100, "true_members")
# extract top network
top_network <- getTopNetwork(population)
}

}
