% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializePopulation.R
\name{initializePopulation}
\alias{initializePopulation}
\title{Initialize a population of \code{PrixFixeNetwork}s}
\usage{
initializePopulation(
  pf_data,
  population_size,
  members = c("true_members", "null_members")
)
}
\arguments{
\item{pf_data}{a \code{PFData} object generated by \code{PFDataLoader}}

\item{population_size}{The number of networks to generate for the population.}

\item{members}{a character string of either \code{true_members} or \code{null_members}.}
}
\value{
A list of size \code{population_size} containing \code{PrixFixeNetwork} objects.
}
\description{
Given a \code{PFData} object, generate a random population of \code{PrixFixeNetwork}
objects of size \code{population_size}.
}
\examples{
\dontrun{
#' # load example PFData (FA genes)
data(PF_FanconiAnemia)
# generate population of subnetworks
population <- initializePopulation(PF_FanconiAnemia, population_size=100, "true_members")
}

}
