% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCFNData.R
\name{loadCFNData}
\alias{loadCFNData}
\title{\code{loadCFNData}}
\usage{
loadCFNData(path)
}
\arguments{
\item{path}{filepath to tab delimited input file containing CFN interactions and weights.}
}
\value{
A formatted tibble with CFN information
}
\description{
Load co-function network data into memory. Called by \code{PFDataLoader}, not intended to be
directly invoked by user.
}
