% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLociData.R
\name{loadLociData}
\alias{loadLociData}
\title{\code{loadLociData}}
\usage{
loadLociData(path)
}
\arguments{
\item{path}{filepath to tab delimited input file containing loci membership.}
}
\value{
A formatted tibble with loci information
}
\description{
Load loci data into memory. Called by \code{PFDataLoader}, not intended to be
directly invoked by user.
}
