% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locus.R
\docType{class}
\name{Locus-class}
\alias{Locus-class}
\title{An S4 class to represent a Locus}
\description{
An S4 class to represent a Locus
}
\section{Slots}{

\describe{
\item{\code{locus_id}}{a character to label the locus (derived from the 2nd column of the loci data).}

\item{\code{true_members}}{a character vector containing the true members of the locus}

\item{\code{null_members}}{a character vector containing bin-matched null members for significance testing.}

\item{\code{cardinality}}{an integer for the length of genes within the locus}
}}

