% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopar_progress_bar.R
\name{make_progress_bar}
\alias{make_progress_bar}
\title{\code{dopar_progress_bar}}
\usage{
make_progress_bar(max)
}
\arguments{
\item{max}{The number of max iterations to pass to \code{txtProgressBar}}
}
\value{
progress bar function
}
\description{
Convenient wrapper for generate a progress bar within a parallel loop. Not
intended to be called by user. Specifically this is a wrapper function to yield
the function for the \code{.combine} operator in \code{foreach}.
}
