% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PF_FanconiAnemia}
\alias{PF_FanconiAnemia}
\title{PFData Object of FA genes}
\format{An object of class \code{PFData} of length 1.}
\usage{
PF_FanconiAnemia
}
\description{
An example dataset generated using the Fanconi Anemia genes and the STRING database.
}
\keyword{datasets}
