% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFData.R
\docType{class}
\name{PFData-class}
\alias{PFData-class}
\title{An S4 class for storing the input dataset for PrixFixeNetwork generation}
\description{
An S4 class for storing the input dataset for PrixFixeNetwork generation
}
\section{Slots}{

\describe{
\item{\code{loci_data_path}}{character path to file containing the loci data}

\item{\code{cfn_data_path}}{character path to file containing the co-function network data}

\item{\code{loci_data}}{list of \code{Locus} objects}

\item{\code{loci_dataframe}}{a tibble for storing the loci data}

\item{\code{cfn_data}}{a tibble for storing the co-function network data}

\item{\code{degree_info}}{a tibble for storing the gene interaction degree information}
}}

