% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFDataLoader.R
\name{PFDataLoader}
\alias{PFDataLoader}
\title{Build a \code{PFData} Object}
\usage{
PFDataLoader(loci_data_path, cfn_data_path, bin = TRUE, n_bins = 128)
}
\arguments{
\item{loci_data_path}{character path for target loci data. This is expected to be a
tab-delimited file where the first column is a project metalabel, the
second column is the column is a descriptor for the loci, and each subsequent
column is a gene associated with the locus. Each row represents another locus
to consider.}

\item{cfn_data_path}{character path for the co-function network}

\item{bin}{A boolean on whether to bin degrees in the co-function network}

\item{n_bins}{Number of bins to separate degrees into (quantile based binning).}
}
\value{
A \code{PFData} object
}
\description{
Loads data from input file paths and constructs a PFNData class object
comprised of a set of Locus class objects and associated data for analysis.
}
\examples{
\dontrun{
pf_data <- PFDataLoader(
"../app/preloaded_data/FA_loci.txt",
"../app/preloaded_data/Day3_STRING.txt")
}

}
