% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrixFixeNetwork.R
\docType{class}
\name{PrixFixeNetwork-class}
\alias{PrixFixeNetwork-class}
\title{An S4 class to represent a PrixFixeNetwork}
\description{
An S4 class to represent a PrixFixeNetwork
}
\section{Slots}{

\describe{
\item{\code{loci}}{character vector of locus_ids}

\item{\code{genes}}{character vector of selected genes for each locus}

\item{\code{adjacency_matrix}}{A matrix containing the adjacency information for the subnetwork}

\item{\code{s_i_star}}{Numeric selection score}
}}

