% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_func.R
\name{read_data_func}
\alias{read_data_func}
\title{\code{read_data_func}}
\usage{
read_data_func(path)
}
\arguments{
\item{path}{filepath to tab delimited input file}
}
\value{
A tibble
}
\description{
Sipmle wrapper function for reading the target loci. Called by
\code{\link{loadLociData}} and `loadCFNData` within `PFDataLoader`, not meant to invoked
by user.
}
