% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_from_bin.R
\name{resample_from_bin}
\alias{resample_from_bin}
\title{\code{resample_from_bin}}
\usage{
resample_from_bin(degree_info, candidate_gene, size = 1)
}
\arguments{
\item{degree_info}{A tibble containing the CFN degree information}

\item{candidate_gene}{The gene that is to be resampled (i.e., the true locus gene)}

\item{size}{the number of genes to sample from the CFN}
}
\value{
resampled genes
}
\description{
Worker function to resample genes from loci to create the null model.
Not intended to be called by user.
}
