% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleLoci.R
\name{sampleLoci}
\alias{sampleLoci}
\title{\code{sampleLoci}}
\usage{
sampleLoci(pf_data, slot = "true_members")
}
\arguments{
\item{pf_data}{A \code{PFData} object}

\item{slot}{character identifying the slot to sample from}
}
\value{
character vector of sampled values
}
\description{
Worker function to sample genes from loci in parallel.
Not intended to be directly implemented by the user.
}
