% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_null_member.R
\name{synthesize_null_member}
\alias{synthesize_null_member}
\title{\code{synthesize_null_member}}
\usage{
synthesize_null_member(locus, pf_data)
}
\arguments{
\item{locus}{A \code{Locus} object}

\item{pf_data}{A \code{PFData} object}
}
\value{
A \code{Locus} object with the null member slot filled
}
\description{
Worker function. Called by \code{synthesizeNullMembers}, not intended to be
called by the user.
}
