% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizeNullMembers.R
\name{synthesizeNullMembers}
\alias{synthesizeNullMembers}
\title{\code{synthesizeNullMembers}}
\usage{
synthesizeNullMembers(pf_data)
}
\arguments{
\item{pf_data}{A \code{PFData} object}
}
\value{
A \code{PFData} object with completed loci_data
}
\description{
Worker function. Called by \code{evaluatePopulationSignificance}, not intended to be
called by the user.
}
